Remove comment filling in markdown mode
I'm not sure what this does, but in my setup at least, this is the cause
of a very annoying bug: the first time a line is auto-wrapped when
typing into a buffer, the last-typed space is truncated when the line
gets wrapped.

In other words, this bug means that every time you open a particular
markdown buffer for the first time in your current Emacs session, the
first time your line is auto-wrapped due to its length, your last-typed
space char is lost.

This has been annoying me for years. Glad I finally found it.