Replace escreen with tab-bar-mode, new in Emacs 27
tab-bar-mode is built-in, and works well with multiple frames. It
allowed me to remove a lot of extra configuration; hooray!

The UI for this tab bar is not yet implemented in the Cocoa build of
Emacs, so I'm using a simple UI I've implemented in the message line,
which I used for escreen.