Magit: preserve the cursor position when refreshing git status
My workflow is that I review the change in the git status view. If I see a line that I don't like, I
go to it and hit enter, edit the line, and go back to git status and hit refresh. I want the cursor
to remain where it was.