Magit: remove directory selection logic
I think I had this code in here so that the top-level git repo would be used for magit-status rather
than a git repo contained within that top repo. Maybe this was for working with git submodules?

In any case, this breaks the UX when working with a mono repo where I have separate projectile
projects within that mono repo. I guess I could've instead improved this by checking if the
projectile directory is itself a git repo, and if not, then don't set the directory for magit
status.